+++ plugins/bulk_upload/drop_folder_xml/lib/api/KalturaDropFolderXmlBulkUploadJobData.php
@@ -3,7 +3,7 @@
 /**
  *
  * Represents the Bulk upload job data for drop folder xml bulk upload
- * @package  plugins.dropFolderXmlBulkUpload
+ * @package plugins.dropFolderXmlBulkUpload
  * @subpackage api.objects
  *
  */